refactor: create a wrapper modal
